Alerting: Rename .test.scenario. files to .scenario.
