+++ packages/grafana-alerting/src/grafana/contactPoints/components/ContactPointSelector/ContactPointSelector.story.tsx
@@ -7,7 +7,7 @@ import { defaultDecorators } from '../../../../../tests/story-utils';
 
 import { ContactPointSelector, ContactPointSelectorProps } from './ContactPointSelector';
 import mdx from './ContactPointSelector.mdx';
-import { simpleContactPointsListScenario, withErrorScenario } from './ContactPointSelector.test.scenario';
+import { simpleContactPointsListScenario, withErrorScenario } from './ContactPointSelector.scenario';
 
 const meta: Meta<typeof ContactPointSelector> = {
   component: ContactPointSelector,
