+++ packages/grafana-alerting/src/grafana/contactPoints/components/ContactPointSelector/ContactPointSelector.test.tsx
@@ -9,7 +9,7 @@ import {
   contactPointsListWithUnusableItemsScenario,
   simpleContactPointsList,
   simpleContactPointsListScenario,
-} from './ContactPointSelector.test.scenario';
+} from './ContactPointSelector.scenario';
 
 const server = setupMockServer();
