+++ packages/grafana-alerting/src/testing.ts
@@ -7,5 +7,5 @@ export * from './grafana/api/notifications/v0alpha1/mocks/fakes/Receivers';
 export * from './grafana/api/notifications/v0alpha1/mocks/fakes/Routes';
 
 // scenarios
-export * from './grafana/contactPoints/components/ContactPointSelector/ContactPointSelector.test.scenario';
-export * from './grafana/notificationPolicies/components/RoutingTreeSelector/RoutingTreeSelector.test.scenario';
+export * from './grafana/contactPoints/components/ContactPointSelector/ContactPointSelector.scenario';
+export * from './grafana/notificationPolicies/components/RoutingTreeSelector/RoutingTreeSelector.scenario';
