+++ public/app/features/alerting/unified/components/rule-viewer/ContactPointLink.scenario.ts
